requirements.txt
amqp==5.1.1
asgiref @ file:///C:/ci_311/asgiref_1676448296195/work
billiard==4.1.0
celery==5.3.1
click==8.1.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
Django @ file:///C:/ci_311/django_1676469589007/work
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
django-timezone-field==5.1
djangorestframework==3.14.0
djangorestframework-gis==1.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
eventlet==0.33.3
GDAL @ file:///C:/Users/sagyan/Desktop/NAXA/GDAL-3.2.3-cp38-cp38-win_amd64.whl#sha256=8f9ba3e6daab17d59d47a1238d734072007245226d167681cd84f5a20d749605
greenlet==2.0.2
iniconfig==2.0.0
kombu==5.3.1
packaging==23.1
phonenumbers==8.13.15
pluggy==1.2.0
prompt-toolkit==3.0.39
psycopg2==2.9.6
psycopg2-binary==2.9.6
psycopg2-pool==1.1
PyJWT==2.7.0
pytest==7.4.0
pytest-django==4.5.2
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.3
redis==4.6.0
six==1.16.0
sqlparse @ file:///C:/ci_311/sqlparse_1676456541280/work
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
